reshuffle section row
